Transfere a criação de tarefas para componente separado
