+++ sarinha-app/src/app/globals.css
@@ -13,6 +13,5 @@
 .container {
   min-height: 100vh;
   min-width: 225px;
-  background-color: "#F5F5F5";
   padding-bottom: 100px;
 }
